설정 (키매핑 제외)

+++ Assets/Scripts/Locale/Translate.cs
@@ -3,16 +3,24 @@ using System.Collections.Generic;
 using System.IO;
 using System.Net;
 using System.Text;
+using Manager;
 using Serialization;
 using UnityEngine;
 using Utils;
 
 namespace Locale {
     public class Translate : MonoBehaviour {
-        public static Dictionary<string, Dictionary<Languages, string>> TranslationsDict;
-        public static List<Languages> AvailableLanguages;
+        public static Dictionary<string, Dictionary<Language, string>> TranslationsDict;
 
-        public static bool TryGet(string key, out string value, Languages? language = null) {
+        public static List<Language> AvailableLanguages {
+            get {
+                if (_availableLanguages == null) LoadLanguages();
+                return _availableLanguages;
+            }
+        }
+        private static List<Language> _availableLanguages;
+
+        public static bool TryGet(string key, out string value, Language? language = null) {
             if (TranslationsDict == null) TranslationsDict = LoadLanguages();
             language ??= Settings.CurrentLanguage;
             if (!TranslationsDict.ContainsKey(key)) {
@@ -23,13 +31,13 @@ namespace Locale {
             var dict2 = TranslationsDict[key];
 
             if (dict2.TryGetValue(language.Value, out value)) return true;
-            return dict2.TryGetValue(Languages.English, out value);
+            return dict2.TryGetValue(Language.English, out value);
         }
 
-        public static string Get(string key, Languages? language = null) =>
+        public static string Get(string key, Language? language = null) =>
             TryGet(key, out string value, language) ? value : $"KeyNotFound {key}";
 
-        public static bool TryGetFormatted(string key, out string value, Languages? language = null,
+        public static bool TryGetFormatted(string key, out string value, Language? language = null,
             params object[] formats) {
             if (formats == null) return TryGet(key, out value, language);
             if (!TryGet(key, out value, language)) return false;
@@ -37,29 +45,30 @@ namespace Locale {
             return true;
         }
 
-        public static string GetFormatted(string key, Languages? language = null, params object[] formats) =>
+        public static string GetFormatted(string key, Language? language = null, params object[] formats) =>
             TryGetFormatted(key, out string value, language, formats) ? value : $"KeyNotFound {key}";
 
-        public static Dictionary<string, Dictionary<Languages, string>> LoadLanguages() {
-            AvailableLanguages = new List<Languages>();
-            
-            #if UNITY_EDITOR
-            var path = Path.Combine(Constants.ResourcePath, "translation.csv");
-            try {
-                using var client = new WebClient();
-                client.DownloadFile(
-                    "https://docs.google.com/spreadsheets/d/14LI14cXLixkUbz1Ap4UzH1hRZkZCKbd6ysrvhExC_Mo/export?format=csv",
-                    path);
-            } catch { }
+        public static Dictionary<string, Dictionary<Language, string>> LoadLanguages() {
+            _availableLanguages = new List<Language>();
+            string result;
+            if (Application.isEditor && Application.isPlaying) {
+                var path = Path.Combine(Constants.ResourcePath, "translation.csv");
+                try {
+                    using var client = new WebClient();
+                    client.DownloadFile(
+                        "https://docs.google.com/spreadsheets/d/14LI14cXLixkUbz1Ap4UzH1hRZkZCKbd6ysrvhExC_Mo/export?format=csv",
+                        path);
+                } catch { }
 
-            File.Delete(path + ".tmp");
-            File.Copy(path, path + ".tmp");
-            path = path + ".tmp";
-            var result = File.ReadAllBytes(path).Encode(Encoding.GetEncoding(65001)).Replace("\r", "");
-            File.Delete(path);
-            #else
-            var result = Resources.Load<TextAsset>("translation").text.Replace("\r", "");
-            #endif
+                File.Delete(path + ".tmp");
+                File.Copy(path, path + ".tmp");
+                path = path + ".tmp";
+                result = File.ReadAllBytes(path).Encode(Encoding.GetEncoding(65001)).Replace("\r", "");
+                File.Delete(path);
+            } else {
+                if (!Application.isPlaying) Debug.Log("Not loading sheet due to not playing ");
+                result = Resources.Load<TextAsset>("translation").text.Replace("\r", "");
+            }
             var csv = new Csv(result);
             var keys = new List<string>();
             var langs = new List<string>();
@@ -74,13 +83,13 @@ namespace Locale {
                 cols.Add(rows);
             }
 
-            var datas = new Dictionary<string, Dictionary<Languages, string>>();
+            var datas = new Dictionary<string, Dictionary<Language, string>>();
             var col = 0;
             foreach (var key in keys) {
-                datas[key] = new Dictionary<Languages, string>();
+                datas[key] = new Dictionary<Language, string>();
                 var row = 0;
                 foreach (var langstr in langs) {
-                    if (Enum.TryParse<Languages>(langstr, out var lang)) {
+                    if (Enum.TryParse<Language>(langstr, out var lang)) {
                         AvailableLanguages.Add(lang);
                         datas[key][lang] = cols[col][row];
                     }
@@ -90,7 +99,7 @@ namespace Locale {
 
                 col++;
             }
-
+            Debug.Log("End loaded");
             return datas;
         }
     }
